refactor destination and edit destination WIP
